pyproject.toml
[project]
name = "itten"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "scipy>=1.17.0",
    "torch>=2.10.0",
    "trio>=0.32.0",
]
